Don't load data at z16
We seem to have memory issues on Android phones. As a temporary fix,
we'll wait until zoom level 17 to load parcels and completed responses,
and we won't buffer as much extra data until zoom 18. This is just
temporary, since it's useful to have parcels/respones at z16 in the
field.